+++ tests/ai_safety/trustyai_service/fairness/conftest.py
@@ -7,7 +7,7 @@ from ocp_resources.namespace import Namespace
 from ocp_resources.secret import Secret
 from ocp_resources.serving_runtime import ServingRuntime
 
-from tests.model_explainability.trustyai_service.trustyai_service_utils import (
+from tests.ai_safety.trustyai_service.trustyai_service_utils import (
     wait_for_isvc_deployment_registered_by_trustyai_service,
 )
 from utilities.constants import ModelFormat, KServeDeploymentType, RuntimeTemplates
